getMaxInputSandboxSize returns long (used to be double)

+++ org.glite.wmproxy/src/server/wmpresponsestruct.h
@@ -99,7 +99,7 @@ struct jobCancelResponse {
 };
 
 struct getMaxInputSandboxSizeResponse {  
-  double size;
+  long size;
 };
 
 struct getSandboxDestURIResponse {
